draw_list: implement modrect on image so we can skip redundant updates

+++ src/glyph.h
@@ -193,16 +193,16 @@ struct text_segment
     std::string language;
     font_face *face;
     int font_size;
-    int x, y;
-    int baseline_shift;
-    int line_spacing;
-    int tracking;
+    float x, y;
+    float baseline_shift;
+    float line_spacing;
+    float tracking;
     uint32_t color;
 
     text_segment() = default;
     text_segment(std::string text, std::string languag);
     text_segment(std::string text, std::string language, font_face *face,
-        int font_size, int x, int y, uint32_t color);
+        int font_size, float x, float y, uint32_t color);
 };
 
 inline text_segment::text_segment(std::string text, std::string language) :
@@ -210,7 +210,7 @@ inline text_segment::text_segment(std::string text, std::string language) :
     x(0), y(0), baseline_shift(0), line_spacing(0), tracking(0), color(0) {}
 
 inline text_segment::text_segment(std::string text, std::string language,
-    font_face *face, int font_size, int x, int y, uint32_t color) :
+    font_face *face, int font_size, float x, float y, uint32_t color) :
     text(text), language(language), face(face), font_size(font_size),
     x(x), y(y), baseline_shift(0), line_spacing(0), tracking(0), color(color) {}
